made data functions more transparent

+++ cryogrid_pytools/outputs.py
@@ -234,7 +234,7 @@ def read_OUT_regridded_files(
             "Depths are the same for all gridcells. Setting depth as the dimension."
         )
         ds = ds.swap_dims(level="depth")
-        ds = ds.reset_coords("elevation")
+        ds = ds.reset_coords("elevation").astype("float32")
 
     return ds
